issue #118 - uppercase constants + minor code improvements

+++ src/main/java/io/proximax/sdk/model/blockchain/BlockInfo.java
@@ -213,7 +213,7 @@ public class BlockInfo {
        return new BlockInfo(dto.getMeta().getHash(),
              dto.getMeta().getGenerationHash(),
              Optional.of(toBigInt(dto.getMeta().getTotalFee())),
-             Optional.of(dto.getMeta().getNumTransactions().intValue()),
+             Optional.of(dto.getMeta().getNumTransactions()),
              dto.getBlock().getSignature(),
              new PublicAccount(dto.getBlock().getSigner(), networkType),
              networkType,
